+++ NEWS.md
@@ -1,6 +1,7 @@
 # version 0.2.2
 
 - Add include_cell argument to exact_extract (Michael Sumner)
+- Fix error thrown by exact_extract when a polygon is partially outside the extent of the value raster but fully within the extent of the weighting raster.
 
 # version 0.2.1